Exception Filters - parte 1

+++ src/users/dto/repositories/users.repository.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateUserDto } from '../create-user.dto';
 import { UpdateUserDto } from '../update-user.dto';
@@ -9,6 +13,18 @@ export class UsersRepository {
   constructor(private readonly prisma: PrismaService) {}
 
   public async create(createUserDto: CreateUserDto): Promise<UserEntity> {
+    const userExist = await this.prisma.user.findUnique({
+      where: {
+        email: createUserDto.email,
+      },
+    });
+
+    if (userExist) {
+      throw new UnauthorizedException(
+        `There is already a user created with the email ${createUserDto.email}`,
+      );
+    }
+
     const user = await this.prisma.user.create({
       data: createUserDto,
     });
@@ -60,6 +76,16 @@ export class UsersRepository {
   }
 
   public async remove(id: number) {
+    const user = await this.prisma.user.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!user) {
+      throw new NotFoundException(`User id ${id} not found`);
+    }
+
     await this.prisma.user.delete({
       where: {
         id,
